requirements.txt
asttokens==3.0.0
click==7.1.2
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
executing==2.1.0
Flask==1.1.2
gunicorn==20.0.4
ipykernel==6.29.5
ipython==8.30.0
itsdangerous==1.1.0
jedi==0.19.2
Jinja2==2.11.3
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==1.1.1
matplotlib-inline==0.1.7
neo4j==5.27.0
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
setuptools==75.6.0
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
wcwidth==0.2.13
Werkzeug==1.0.1
